Allocate reservations memory in heap instead of stack
In the elf_load we need to allocate memory for the reservations.
In the old code this memory was allocated on the stack and was hence
freed after the loading was done.

This meant that the vspace.reservation LL was not accurate later on.

We now allocate that on the heap so that we do no have to free that
memory.